# sdks/python/.speakeasy/out.openapi.yaml
openapi: "3.1.0"
info:
  title: "VulkanAPI"
  version: "0.1.0"
paths:
  "/components/":
    get:
      tags:
        - "components"
      summary: "List Components"
      description: "List all components."
      operationId: "list_components"
      parameters:
        - name: "include_archived"
          in: "query"
          required: false
          schema:
            type: "boolean"
            default: false
            title: "Include Archived"
        - name: "project_id"
          in: "query"
          required: false
          schema:
            anyOf:
              - type: "string"
              - type: "null"
            title: "Project Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema:
                type: "array"
                items: {"$ref": "#/components/schemas/Component"}
                title: "Response List Components"
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "list"
    post:
      tags:
        - "components"
      summary: "Create Component"
      description: "Create a new component."
      operationId: "create_component"
      parameters:
        - name: "project_id"
          in: "query"
          required: false
          schema:
            anyOf:
              - type: "string"
              - type: "null"
            title: "Project Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema: {"$ref": "#/components/schemas/ComponentUpdate"}
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/Component"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "create"
  "/components/{component_name}":
    get:
      tags:
        - "components"
      summary: "Get Component"
      description: "Get a component by name."
      operationId: "get_component"
      parameters:
        - name: "component_name"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Component Name"
        - name: "project_id"
          in: "query"
          required: false
          schema:
            anyOf:
              - type: "string"
              - type: "null"
            title: "Project Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/Component"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
    put:
      tags:
        - "components"
      summary: "Update Component"
      description: "Update a component by name."
      operationId: "update_component"
      parameters:
        - name: "component_name"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Component Name"
        - name: "project_id"
          in: "query"
          required: false
          schema:
            anyOf:
              - type: "string"
              - type: "null"
            title: "Project Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema: {"$ref": "#/components/schemas/ComponentUpdate"}
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/Component"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
    delete:
      tags:
        - "components"
      summary: "Delete Component"
      description: "Delete (archive) a component."
      operationId: "delete_component"
      parameters:
        - name: "component_name"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Component Name"
        - name: "project_id"
          in: "query"
          required: false
          schema:
            anyOf:
              - type: "string"
              - type: "null"
            title: "Project Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
...
  "/internal/runs/{run_id}":
    put:
      tags:
        - "internal"
      summary: "Update Run"
      description: "Update run status and optionally trigger shadow runs."
      operationId: "update_run"
      parameters:
        - name: "run_id"
          in: "path"
          required: true
          schema:
            type: "string"
            format: "uuid"
            title: "Run Id"
      requestBody:
        required: true
        content:
          "application/json":
            schema: {"$ref": "#/components/schemas/Body_update_run"}
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/Run"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "updateRun"
...
  "/runs/{run_id}/data":
    get:
      tags:
        - "runs"
      summary: "Get Run Data"
      description: "Get run data including step outputs and metadata."
      operationId: "get_run_data"
      parameters:
        - name: "run_id"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Run Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/RunData"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "getData"
  "/runs/{run_id}/logs":
    get:
      tags:
        - "runs"
      summary: "Get Run Logs"
      description: "Get logs for a run."
      operationId: "get_run_logs"
      parameters:
        - name: "run_id"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Run Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/RunLogs"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "getLogs"
  "/runs/{run_id}":
    get:
      tags:
        - "runs"
      summary: "Get Run"
      description: "Get run details."
      operationId: "get_run"
      parameters:
        - name: "run_id"
          in: "path"
          required: true
          schema:
            type: "string"
            title: "Run Id"
      responses:
        "200":
          description: "Successful Response"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/Run"}
        "404":
          description: "Not found"
        "422":
          description: "Validation Error"
          content:
            "application/json":
              schema: {"$ref": "#/components/schemas/HTTPValidationError"}
      x-speakeasy-name-override: "get"
...
components:
  schemas:
...
    "Body_update_run":
      properties:
        "status":
          type: "string"
          title: "Status"
        "result":
          type: "string"
          title: "Result"
        "metadata":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Metadata"
      type: "object"
      required:
        - "status"
        - "result"
      title: "Body_update_run"
...
    "Component":
      properties:
        "name":
          type: "string"
          title: "Name"
        "description":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Description"
        "icon":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Icon"
        "component_id":
          type: "string"
          format: "uuid"
          title: "Component Id"
        "archived":
          type: "boolean"
          title: "Archived"
        "created_at":
          type: "string"
          format: "date-time"
          title: "Created At"
        "last_updated_at":
          type: "string"
          format: "date-time"
          title: "Last Updated At"
        "workflow":
          anyOf:
            - {"$ref": "#/components/schemas/Workflow"}
            - type: "null"
      type: "object"
      required:
        - "name"
        - "component_id"
        - "archived"
        - "created_at"
        - "last_updated_at"
      title: "Component"
    "ComponentUpdate":
      properties:
        "name":
          type: "string"
          title: "Name"
        "description":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Description"
        "icon":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Icon"
        "workflow":
          anyOf:
            - {"$ref": "#/components/schemas/WorkflowBase"}
            - type: "null"
      type: "object"
      required:
        - "name"
      title: "ComponentUpdate"
...
    "DagsterLogDetails":
      properties:
        "log_type":
          type: "string"
          title: "Log Type"
        "message":
          type: "string"
          title: "Message"
      type: "object"
      required:
        - "log_type"
        - "message"
      title: "DagsterLogDetails"
...
    "DependencyDict":
      properties:
        "node":
          type: "string"
          title: "Node"
        "output":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Output"
        "key":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Key"
        "hierarchy":
          anyOf:
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Hierarchy"
      type: "object"
      required:
        - "node"
        - "output"
        - "key"
        - "hierarchy"
      title: "DependencyDict"
      description: "Dict representation of a Dependency object."
...
    "HTTPValidationError":
      properties:
        "detail":
          items: {"$ref": "#/components/schemas/ValidationError"}
          type: "array"
          title: "Detail"
      type: "object"
      title: "HTTPValidationError"
...
    "LogEntry":
      properties:
        "timestamp":
          type: "string"
          format: "date-time"
          title: "Timestamp"
        "step_key":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Step Key"
        "source":
          type: "string"
          title: "Source"
        "event":
          anyOf:
            - {"$ref": "#/components/schemas/UserLogDetails"}
            - {"$ref": "#/components/schemas/DagsterLogDetails"}
          title: "Event"
      type: "object"
      required:
        - "timestamp"
        - "step_key"
        - "source"
        - "event"
      title: "LogEntry"
    "NodeDefinitionDict":
      properties:
        "name":
          type: "string"
          title: "Name"
        "node_type":
          type: "string"
          title: "Node Type"
        "dependencies":
          anyOf:
            - additionalProperties: {"$ref": "#/components/schemas/DependencyDict"}
              type: "object"
            - type: "null"
          title: "Dependencies"
        "metadata":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Metadata"
        "description":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Description"
        "hierarchy":
          anyOf:
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Hierarchy"
      type: "object"
      required:
        - "name"
        - "node_type"
      title: "NodeDefinitionDict"
      description: "Dict representation of a NodeDefinition object."
...
    "PolicyDefinitionDict":
      properties:
        "nodes":
          items: {"$ref": "#/components/schemas/NodeDefinitionDict"}
          type: "array"
          title: "Nodes"
        "input_schema":
          additionalProperties:
            type: "string"
          type: "object"
          title: "Input Schema"
        "config_variables":
          anyOf:
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Config Variables"
        "output_callback":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Output Callback"
      type: "object"
      required:
        - "nodes"
        - "input_schema"
      title: "PolicyDefinitionDict"
      description: "Dict representation of a PolicyDefinition object."
...
    "Run":
      properties:
        "run_id":
          type: "string"
          format: "uuid"
          title: "Run Id"
        "run_group_id":
          anyOf:
            - type: "string"
              format: "uuid"
            - type: "null"
          title: "Run Group Id"
        "policy_version_id":
          type: "string"
          format: "uuid"
          title: "Policy Version Id"
        "status": {"$ref": "#/components/schemas/RunStatus"}
        "result":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Result"
        "input_data":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Input Data"
        "run_metadata":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Run Metadata"
        "created_at":
          type: "string"
          format: "date-time"
          title: "Created At"
        "started_at":
          anyOf:
            - type: "string"
              format: "date-time"
            - type: "null"
          title: "Started At"
        "last_updated_at":
          type: "string"
          format: "date-time"
          title: "Last Updated At"
      type: "object"
      required:
        - "run_id"
        - "policy_version_id"
        - "status"
        - "created_at"
        - "last_updated_at"
      title: "Run"
...
    "RunData":
      properties:
        "run_id":
          type: "string"
          format: "uuid"
          title: "Run Id"
        "run_group_id":
          anyOf:
            - type: "string"
              format: "uuid"
            - type: "null"
          title: "Run Group Id"
        "policy_version_id":
          type: "string"
          format: "uuid"
          title: "Policy Version Id"
        "status": {"$ref": "#/components/schemas/RunStatus"}
        "result":
          anyOf:
            - type: "string"
            - type: "null"
          title: "Result"
        "input_data":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Input Data"
        "run_metadata":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Run Metadata"
        "created_at":
          type: "string"
          format: "date-time"
          title: "Created At"
        "started_at":
          anyOf:
            - type: "string"
              format: "date-time"
            - type: "null"
          title: "Started At"
        "last_updated_at":
          type: "string"
          format: "date-time"
          title: "Last Updated At"
        "steps":
          additionalProperties: {"$ref": "#/components/schemas/StepDetails"}
          type: "object"
          title: "Steps"
      type: "object"
      required:
        - "run_id"
        - "policy_version_id"
        - "status"
        - "created_at"
        - "last_updated_at"
      title: "RunData"
...
    "RunLogs":
      properties:
        "run_id":
          type: "string"
          format: "uuid"
          title: "Run Id"
        "status":
          type: "string"
          title: "Status"
        "last_updated_at":
          type: "string"
          format: "date-time"
          title: "Last Updated At"
        "logs":
          items: {"$ref": "#/components/schemas/LogEntry"}
          type: "array"
          title: "Logs"
      type: "object"
      required:
        - "run_id"
        - "status"
        - "last_updated_at"
        - "logs"
      title: "RunLogs"
...
    "RunStatus":
      type: "string"
      enum:
        - "PENDING"
        - "STARTED"
        - "SUCCESS"
        - "FAILURE"
      title: "RunStatus"
...
    "StepDetails":
      properties:
        "output":
          anyOf:
            - type: "string"
              format: "binary"
            - type: "string"
            - additionalProperties: true
              type: "object"
            - items: {}
              type: "array"
            - type: "integer"
            - type: "number"
            - type: "null"
          title: "Output"
        "metadata":
          anyOf:
            - {"$ref": "#/components/schemas/StepMetadataBase"}
            - type: "null"
      type: "object"
      required:
        - "output"
        - "metadata"
      title: "StepDetails"
    "StepMetadataBase":
      properties:
        "step_name":
          type: "string"
          title: "Step Name"
        "node_type":
          type: "string"
          title: "Node Type"
        "start_time":
          type: "number"
          title: "Start Time"
        "end_time":
          type: "number"
          title: "End Time"
        "error":
          anyOf:
            - type: "string"
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Error"
        "extra":
          anyOf:
            - additionalProperties: true
              type: "object"
            - type: "null"
          title: "Extra"
      type: "object"
      required:
        - "step_name"
        - "node_type"
        - "start_time"
        - "end_time"
      title: "StepMetadataBase"
    "UIMetadata":
      properties:
        "position": {"$ref": "#/components/schemas/UINodePosition"}
        "width":
          type: "number"
          title: "Width"
        "height":
          type: "number"
          title: "Height"
      type: "object"
      required:
        - "position"
        - "width"
        - "height"
      title: "UIMetadata"
    "UINodePosition":
      properties:
        "x":
          type: "number"
          title: "X"
        "y":
          type: "number"
          title: "Y"
      type: "object"
      required:
        - "x"
        - "y"
      title: "UINodePosition"
    "UserLogDetails":
      properties:
        "log_type":
          type: "string"
          title: "Log Type"
          default: "LOG"
        "message":
          type: "string"
          title: "Message"
        "level":
          type: "string"
          title: "Level"
      type: "object"
      required:
        - "message"
        - "level"
      title: "UserLogDetails"
    "ValidationError":
      properties:
        "loc":
          items:
            anyOf:
              - type: "string"
              - type: "integer"
          type: "array"
          title: "Location"
        "msg":
          type: "string"
          title: "Message"
        "type":
          type: "string"
          title: "Error Type"
      type: "object"
      required:
        - "loc"
        - "msg"
        - "type"
      title: "ValidationError"
    "Workflow":
      properties:
        "spec": {"$ref": "#/components/schemas/PolicyDefinitionDict"}
        "requirements":
          items:
            type: "string"
          type: "array"
          title: "Requirements"
        "variables":
          anyOf:
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Variables"
        "ui_metadata":
          anyOf:
            - additionalProperties: {"$ref": "#/components/schemas/UIMetadata"}
              type: "object"
            - type: "null"
          title: "Ui Metadata"
        "workflow_id":
          type: "string"
          format: "uuid"
          title: "Workflow Id"
        "status": {"$ref": "#/components/schemas/WorkflowStatus"}
      type: "object"
      required:
        - "spec"
        - "requirements"
        - "workflow_id"
        - "status"
      title: "Workflow"
    "WorkflowBase":
      properties:
        "spec": {"$ref": "#/components/schemas/PolicyDefinitionDict"}
        "requirements":
          items:
            type: "string"
          type: "array"
          title: "Requirements"
        "variables":
          anyOf:
            - items:
                type: "string"
              type: "array"
            - type: "null"
          title: "Variables"
        "ui_metadata":
          anyOf:
            - additionalProperties: {"$ref": "#/components/schemas/UIMetadata"}
              type: "object"
            - type: "null"
          title: "Ui Metadata"
      type: "object"
      required:
        - "spec"
        - "requirements"
      title: "WorkflowBase"
    "WorkflowStatus":
      type: "string"
      enum:
        - "VALID"
        - "INVALID"
      title: "WorkflowStatus"